5.6

+++ User/hardware.h
@@ -10,8 +10,8 @@
 
 double BMYBJ_YB(u16 init_freq, u8 init_temp, u16 freq, u8 temp);
 double MRYBJ_YB(u16 init_freq, u8 init_temp, u16 freq, u8 temp);
-double GJJ_YL(u16 init_freq, u8 init_temp, u16 freq, u8 temp, u16 K);
-double TYLH_YL_MPa(u16 init_freq, u8 init_temp, u16 freq, u8 temp, u16 K);
-double SYJ_YL_MPa(u16 init_freq, u8 init_temp, u16 freq, u8 temp, u16 A, u16 B, u16 C, u16 b);
-double MSJ_YL_KN(u16* init_freq, u8 init_temp, u16* freq, u8 temp, u8 size, u16 a, u16 b);
+double GJJ_YL(u16 init_freq, u8 init_temp, u16 freq, u8 temp, int32_t K);
+double TYLH_YL_MPa(u16 init_freq, u8 init_temp, u16 freq, u8 temp, int32_t K);
+double SYJ_YL_MPa(u16 init_freq, u8 init_temp, u16 freq, u8 temp, int32_t A, int32_t B, int32_t C, int32_t b);
+double MSJ_YL_KN(u16* init_freq, u8 init_temp, u16* freq, u8 temp, u8 size, int32_t a, int32_t b);
 #endif //ISD2190_HARDWARE_H
